refactor: 북마크 기능 `unused`
