+++ components/hospital/icu/main/chart/selected-chart/chart-header/header-center/header-center.tsx
@@ -1,4 +1,3 @@
-import BookmarkDialog from '@/components/hospital/icu/main/chart/selected-chart/chart-header/header-center/bookmark-dialog'
 import UpdatePatientDialog from '@/components/hospital/icu/main/chart/selected-chart/chart-header/header-center/update-patient-dialog'
 import VitalChartDialog from '@/components/hospital/icu/main/chart/selected-chart/chart-header/header-center/vital-chart/vital-chart-dialog'
 import type { SelectedChart } from '@/types/icu/chart'
@@ -12,11 +11,11 @@ export default function HeaderCenter({
 
   return (
     <div className="flex h-12 items-center justify-center gap-0.5">
-      <BookmarkDialog
+      {/* <BookmarkDialog
         icuChartId={chartData.icu_chart_id}
         bookmarkData={chartData.template}
         icon="star"
-      />
+      /> */}
       <UpdatePatientDialog
         patientData={{ ...patient, isIcu: true }}
         weight={weight}
