+++ unused/bookmark/bookmarks.tsx
@@ -1,7 +1,7 @@
 'use client'
 
 import PreviewDialog from '@/components/hospital/icu/common-dialogs/preview/preview-dialog'
-import { bookmarkColumns } from '@/components/hospital/icu/main/bookmark/bookmark-columns'
+import { bookmarkColumns } from '@/unused/bookmark/bookmark-columns'
 import DataTable from '@/components/ui/data-table'
 import { usePreviewDialogStore } from '@/lib/store/icu/preview-dialog'
 import type { TemplateChart } from '@/types/icu/template'
